[Gluon][AMD] Move gfx1250 APIs to cdna5
