+++ python/triton/experimental/gluon/amd/__init__.py
@@ -1,4 +1,3 @@
-from . import gfx1250
-from . import cdna5
+from . import cdna5, gfx1250
 
-__all__ = ["gfx1250", "cdna5"]
+__all__ = ["cdna5", "gfx1250"]
