+++ python/triton/experimental/gluon/amd/gfx1250/__init__.py
@@ -1,0 +1,2 @@
+from ..cdna5 import *  # noqa: F403
+from ..cdna5 import __all__ as __all__
