+++ python/triton/experimental/gluon/language/amd/__init__.py
@@ -1,9 +1,8 @@
 from .._core import builtin
 from ._layouts import AMDMFMALayout, AMDWMMALayout
-from . import cdna3, cdna4
+from . import cdna3, cdna4, cdna5
 from . import rdna3, rdna4
 from . import gfx1250
-from . import cdna5
 from .slice import slice
 from .warp_pipeline import warp_pipeline_stage
